# services/conduit-core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The payment API for Conduit, a test-mode payments infrastructure platform. Mirrors Stripe's own object naming (payment_intents) and auth model (sk_test_/pk_test_ key pairs) deliberately.",
        "title": "Conduit API",
        "contact": {
            "name": "Conduit"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/payment_intents": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lists the authenticated merchant's payment intents, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment_intents"
                ],
                "summary": "List payment intents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max results (default 50, capped at 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_paymentintent.PaymentIntent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new payment intent in the `created` state. Requires an Idempotency-Key — a duplicate key returns the original response instead of creating a second record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment_intents"
                ],
                "summary": "Create a payment intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payment intent details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_paymentintent.createRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_paymentintent.PaymentIntent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payment_intents/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches a payment intent by id, scoped to the authenticated merchant. Returns 404 for another merchant's payment intent — never a 403 that would confirm it exists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment_intents"
                ],
                "summary": "Get a payment intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment intent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_paymentintent.PaymentIntent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payment_intents/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Drives created|pending -\u003e succeeded|failed. Scores the charge with conduit-risk synchronously first; a decline transitions straight to failed with no ledger entry ever created. An already-terminal payment intent is returned as-is rather than reprocessed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment_intents"
                ],
                "summary": "Confirm a payment intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment intent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_paymentintent.PaymentIntent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payment_intents/{id}/refund": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Drives succeeded -\u003e refunded, posting a second balanced ledger transaction with entries reversed relative to the original charge. Only a succeeded payment intent can be refunded.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment_intents"
                ],
                "summary": "Refund a payment intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment intent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_paymentintent.PaymentIntent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_paymentintent.PaymentIntent": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "failure_reason": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/internal_paymentintent.Status"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_paymentintent.Status": {
            "type": "string",
            "enum": [
                "created",
                "pending",
                "succeeded",
                "failed",
                "refunded"
            ],
            "x-enum-varnames": [
                "StatusCreated",
                "StatusPending",
                "StatusSucceeded",
                "StatusFailed",
                "StatusRefunded"
            ]
        },
        "internal_paymentintent.createRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Merchant secret key, sent as \"Bearer sk_test_...\". Obtained once from POST /v1/merchants.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}